TIFT-Backend-main/TIFT-Backend-main/requirements.txt
Django==5.1.3
aiogram==3.20.0.post0
aiogram-calendar==0.6.0
psycopg2-binary
djangorestframework==3.15.2
drf-yasg==1.21.8
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
requests==2.32.3
django-cors-headers==4.6.0
django-apscheduler==0.7.0
boto3==1.35.55
django-storages==1.14.4
whitenoise==6.8.2
pandas==2.2.3
openpyxl==3.1.5
django-modeltranslation==0.19.14
pillow==11.2.1
xlrd>=2.0.1
weasyprint==65.1
python-docx==1.1.2
reportlab==4.4.1
PyPDF2==3.0.1
qrcode==8.2
django-jazzmin==3.0.1
django-admin-rangefilter==0.13.2
PyMuPDF==1.26.3